+++ programs/cardinal-group-reward-distributor/src/instructions/close_group_reward_distributor.rs
@@ -26,7 +26,7 @@ pub fn handler<'key, 'accounts, 'remaining, 'info>(ctx: Context<'key, 'accounts,
 
     let remaining_accs = &mut ctx.remaining_accounts.iter();
     match group_reward_distributor.reward_kind {
-        k if k == GroupRewardDistributorKind::Mint as u8 => {
+        k if k == GroupRewardDistributorKind::Mint => {
             let cpi_accounts = SetAuthority {
                 account_or_mint: ctx.accounts.reward_mint.to_account_info(),
                 current_authority: group_reward_distributor.to_account_info(),
@@ -35,7 +35,7 @@ pub fn handler<'key, 'accounts, 'remaining, 'info>(ctx: Context<'key, 'accounts,
             let cpi_context = CpiContext::new(cpi_program, cpi_accounts).with_signer(group_reward_distributor_signer);
             token::set_authority(cpi_context, AuthorityType::MintTokens, Some(ctx.accounts.reward_mint.key()))?;
         }
-        k if k == GroupRewardDistributorKind::Treasury as u8 => {
+        k if k == GroupRewardDistributorKind::Treasury => {
             let group_reward_distributor_token_account_info = next_account_info(remaining_accs)?;
             let group_reward_distributor_token_account = Account::<TokenAccount>::try_from(group_reward_distributor_token_account_info)?;
             let authority_token_account_info = next_account_info(remaining_accs)?;
